+++ packages/kmp-compose/src/commonMain/kotlin/ly/neptune/signal/Signal.kt
@@ -1,8 +1,7 @@
 package ly.neptune.signal
 
 object Signal {
-    const val StandardVersion = "0.1.0"
-    const val SdkVersion = "0.1.0"
+    const val StandardVersion = "0.1.1"
+    const val SdkVersion = "0.1.1"
     const val Name = "Neptune. Signal"
 }
-